// src/types/enums.ts
// Enums
export enum ApplicationMsgDirection {
  FromStudents = 0,
  ToStudents = 1
}

export enum ApplicationStatus {
  InProgress = 0,
  New = 1,
  MeetPlanned = 2,
  Accepted = 3,
  Rejected = 4
}

export enum CaseReactionType {
  Neutral = 0,
  Positive = 1,
  Negative = 2
}

export enum DayOfWeek {
  Monday = 0,
  Tuesday = 1,
  Wednesday = 2,
  Thursday = 3,
  Friday = 4,
  Saturday = 5,
  Sunday = 6
}

export enum ProjectStatus {
  Created = 0,
  InWork = 1,
  Completed = 2
}

export enum Semester {
  Autumn = 0,
  Spring = 1
}